Change temperature to a double, add limited error handling, add requests, fix paths

+++ db/migrations/001_init.sql
@@ -1,5 +1,5 @@
 CREATE TABLE temperature (
 	id serial PRIMARY KEY,
 	timestamp TIMESTAMP NOT NULL,
-	temperature INTEGER
+	temperature REAL
 );
